feat: Create JoobleScraper
The next URL is not working yet.
The paege is loading as the mobile version, so it differes with what I see on the browser.
Make request in a way that shows web version.

https://au.jooble.org/jobs-working-holiday-visa?p=1
redirects to
https://au.jooble.org/m/jobs-working-holiday-visa?p=1

That is why I was seeing inconsistency in the number of results, 14 instead of 19.